requirements.txt
appdirs==1.4.4
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
click==7.1.2
cloudscraper==1.2.71
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
pixi==1.0.1
pixiv-api==0.3.7
pyparsing==3.1.2
requests==2.32.3
requests-toolbelt==1.0.0
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2